Moves decoder and encoder tests to shared
